+++ src/domain/scenes/card-selector/card-selector-cards-list/card-selector-cards-list.h
@@ -8,17 +8,17 @@ namespace Game
   struct CardSelectorCardsListConfig
   {
     const Vector position;
-    const float width;
-    const float height;
+    const int width;
+    const int height;
   };
 
   class CardSelectorCardsList
   {
   private:
     const Vector position;
-    const float width;
-    const float height;
-    const float cardSize;
+    const int width;
+    const int height;
+    const int cardSize;
     const Vector initialCardPosition;
 
   public:
